pyproject.toml
[tool.poetry]
name = "poker"
version = "1.0a"
description = "Backend for playing poker via websockets."
authors = ["raindinners"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.9"
colorlog = "^6.8.2"
stringcase = "^1.2.0"
pokerengine = "^1.8"
apscheduler = "^3.10.4"
pydantic = "^2.3.0"
pydantic-settings = "^2.0.3"
fastapi = "^0.111.1"
fastapi-distributed-websocket = "^0.2.0"
uvicorn = "^0.30.1"
redis = "^5.0.4"
sqlalchemy = "^2.0.7"
alembic = "^1.10.2"
asyncpg = "^0.29.0"
psycopg2 = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.2"
black = "^24.4.2"
ruff = "^0.5.0"
isort = "^5.12.0"
mypy = "^1.2.0"

[tool.ruff]
line-length = 99
select = [
    "C4",
    "E",
    "F",
    "T10",
    "T20",
    "Q",
    "RET",
]
ignore = [
    "E501",
]
exclude = [
    ".git",
    ".venv",
    "logs",
]
force-exclude = true

[tool.black]
line-length = 99
target-version = ['py38', 'py39', 'py310', 'py311']
exclude = '''
(
    \.git
    \.ruff_cache
    \.venv
  | logs
)
'''

[tool.isort]
profile = "black"
src_paths = [
    "app",
    "tests",
]

[tool.mypy]
plugins = [
    "pydantic.mypy",
]
disable_error_code = [
    "import",
]
show_error_codes = true
show_error_context = true
namespace_packages = true

[[tool.mypy.overrides]]
module = [
    "passlib",
]
ignore_missing_imports = true
disallow_untyped_defs = true
